requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
PyMySQL==1.1.0

# Security and JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Redis for caching
redis==5.0.1

# Data validation (built into FastAPI via Pydantic)
email-validator==2.1.0

# Production server (uvicorn with gunicorn)
gunicorn==21.2.0

# Templates
jinja2==3.1.2